+++ tests/dummy/app/controllers/components-examples/flexberry-toggler/settings-example-inner.js
@@ -58,6 +58,15 @@ export default Ember.Controller.extend({
   iconClass: '',
 
   /**
+    Text for inner 'flexberry-togggler' component 'duration' property.
+
+    @property duration
+    @type Number
+    @default 350
+  */
+    duration: 350,
+
+  /**
     Is accordion expanded?
 
     @property expanded